avanze en buscar madera

+++ src/Characters/Enemigo.java
@@ -2,7 +2,7 @@ package Characters;
 
 public class Enemigo extends Personaje {
 
-    public Enemigo(String nombre, int vida, int damage, int resistencia, double velocidad){
+    public Enemigo(String nombre, int vida, int damage, int resistencia, float velocidad){
         super(nombre, vida,damage,resistencia,velocidad);
     }
     public Enemigo(){
